requirements.txt
supabase
pinecone
openai
python-dotenv
